+++ src/main/scala/com/github/unchama/seichiassist/subsystems/autosave/System.scala
@@ -3,9 +3,8 @@ package com.github.unchama.seichiassist.subsystems.autosave
 import cats.effect.{Sync, Timer}
 import com.github.unchama.seichiassist.concurrent.PluginExecutionContexts
 import com.github.unchama.seichiassist.meta.subsystem.StatefulSubsystem
-import com.github.unchama.seichiassist.subsystems.autosave.application.{CanNotifySaves, CanSaveWorlds, SystemConfiguration}
+import com.github.unchama.seichiassist.subsystems.autosave.application.{CanNotifySaves, CanSaveWorlds, SystemConfiguration, WorldSaveRoutine}
 import com.github.unchama.seichiassist.subsystems.autosave.bukkit.instances.{SyncCanNotifyBukkitSaves, SyncCanSaveBukkitWorlds}
-import com.github.unchama.seichiassist.subsystems.autosave.bukkit.task.WorldSaveRoutine
 
 object System {
   def wired[
@@ -15,10 +14,9 @@ object System {
     import PluginExecutionContexts._
 
     implicit val _configuration: SystemConfiguration = configuration
-    implicit val _instances: CanNotifySaves[F] with CanSaveWorlds[F] =
-      new SyncCanNotifyBukkitSaves[F] with SyncCanSaveBukkitWorlds[F] {
-        override val F: Sync[F] = implicitly
-      }
+
+    implicit val _canSaveWorlds: CanSaveWorlds[F] = SyncCanSaveBukkitWorlds[F]
+    implicit val _canNotifySaves: CanNotifySaves[F] = SyncCanNotifyBukkitSaves[F]
 
     val repeatedJobs = List[F[Nothing]](
       WorldSaveRoutine()
