+++ src/main/scala/com/github/unchama/seichiassist/subsystems/autosave/application/WorldSaveRoutine.scala
@@ -1,9 +1,8 @@
-package com.github.unchama.seichiassist.subsystems.autosave.bukkit.task
+package com.github.unchama.seichiassist.subsystems.autosave.application
 
 import cats.Applicative
 import cats.effect.{Sync, Timer}
 import com.github.unchama.concurrent.{RepeatingRoutine, RepeatingTaskContext}
-import com.github.unchama.seichiassist.subsystems.autosave.application.{CanNotifySaves, CanSaveWorlds, SystemConfiguration}
 import org.bukkit.ChatColor.AQUA
 
 object WorldSaveRoutine {
