+++ src/main/scala/com/github/unchama/seichiassist/subsystems/autosave/bukkit/instances/SyncCanNotifyBukkitSaves.scala
@@ -5,11 +5,9 @@ import com.github.unchama.seichiassist.subsystems.autosave.application.CanNotify
 import com.github.unchama.seichiassist.util.Util
 import org.bukkit.Bukkit
 
-trait SyncCanNotifyBukkitSaves[F[_]] extends CanNotifySaves[F] {
+object SyncCanNotifyBukkitSaves {
 
-  val F: Sync[F]
-
-  final override def notify(message: String): F[Unit] = F.delay {
+  def apply[F[_] : Sync]: CanNotifySaves[F] = (message: String) => Sync[F].delay {
     Util.sendEveryMessage(message)
     Bukkit.getLogger.info(message)
   }
