+++ src/main/scala/com/github/unchama/seichiassist/subsystems/autosave/bukkit/instances/SyncCanSaveBukkitWorlds.scala
@@ -7,41 +7,41 @@ import org.bukkit.{Bukkit, World}
 import java.lang.reflect.Field
 import scala.annotation.tailrec
 
-trait SyncCanSaveBukkitWorlds[F[_]] extends CanSaveWorlds[F] {
+object SyncCanSaveBukkitWorlds {
 
-  val F: Sync[F]
+  def apply[F[_] : Sync]: CanSaveWorlds[F] = new CanSaveWorlds[F] {
+    override val saveAllWorlds: F[Unit] = Sync[F].delay {
+      def saveWorld(world: World): Unit = {
+        // WARNを防ぐためMinecraftサーバーデフォルトの自動セーブは無効化
+        val server = getFieldAsAccessibleField(Bukkit.getServer.getClass, "console")
+          .getOrElse(return)
+          .get(Bukkit.getServer)
 
-  final override val saveAllWorlds: F[Unit] = F.delay {
-    def saveWorld(world: World): Unit = {
-      // WARNを防ぐためMinecraftサーバーデフォルトの自動セーブは無効化
-      val server = getFieldAsAccessibleField(Bukkit.getServer.getClass, "console")
-        .getOrElse(return)
-        .get(Bukkit.getServer)
+        getFieldAsAccessibleField(server.getClass, "autosavePeriod")
+          .getOrElse(return)
+          .set(server, 0)
 
-      getFieldAsAccessibleField(server.getClass, "autosavePeriod")
-        .getOrElse(return)
-        .set(server, 0)
-
-      world.save()
-    }
-
-    @tailrec
-    def getFieldAsAccessibleField(clazz: Class[_], name: String): Option[Field] = {
-      clazz.getDeclaredFields.find(_.getName.equals(name)) match {
-        case s@Some(field) =>
-          field.setAccessible(true)
-          s
-        case None =>
-          clazz.getSuperclass match {
-            case null => None
-            case s => getFieldAsAccessibleField(s, name)
-          }
+        world.save()
       }
+
+      @tailrec
+      def getFieldAsAccessibleField(clazz: Class[_], name: String): Option[Field] = {
+        clazz.getDeclaredFields.find(_.getName.equals(name)) match {
+          case s@Some(field) =>
+            field.setAccessible(true)
+            s
+          case None =>
+            clazz.getSuperclass match {
+              case null => None
+              case s => getFieldAsAccessibleField(s, name)
+            }
+        }
+      }
+
+      import scala.jdk.CollectionConverters._
+
+      Bukkit.getServer.getWorlds.asScala.foreach(saveWorld)
     }
-
-    import scala.jdk.CollectionConverters._
-
-    Bukkit.getServer.getWorlds.asScala.foreach(saveWorld)
   }
 
 }
